pyproject.toml
[project]
name = "youtube-subs"
version = "0.1.0"
description = ""
authors = [
    {name = "Daniel Paurat",email = "daniel.paurat@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<3.15"
dependencies = [
    "youtube-transcript-api (>=1.2.3,<2.0.0)",
    "ollama (>=0.6.0,<0.7.0)",
    "pyreadline3 (>=3.5.4,<4.0.0)",
    "requests (>=2.31.0,<3.0.0)",
    "beautifulsoup4 (>=4.12.2,<5.0.0)",
    "pytest (>=9.0.2,<10.0.0)",
    "rich (>=14.2.0,<15.0.0)",
    "python-dotenv (>=1.0.0,<2.0.0)"
]

[tool.black]
line-length = 120

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
